+++ packages/tools/components-package/nps.js
@@ -11,7 +11,7 @@ const getScripts = (options) => {
 	const createIllustrationsJSImportsScript = illustrations.join(" && ");
 
 	// The script creates the "src/generated/js-imports/Illustration.js" file that registers loaders (dynamic JS imports) for each illustration
-    const createIllustrationsLoadersScript = illustrationsData.map(illustrations => `node ${LIB}/generate-js-imports/illustrations.js ${illustrations.destinationPath} ${illustrations.dynamicImports.outputFile} ${illustrations.set} ${illustrations.collection} ${illustrations.dynamicImports.location} ${illustrations.dynamicImports.filterOut.join(" ")}`).join(" && ");
+    const createIllustrationsLoadersScript = illustrationsData.map(illustrations => `node ${LIB}/generate-js-imports/illustration-collections.js ${illustrations.destinationPath} ${illustrations.dynamicImports.outputFile} ${illustrations.set} ${illustrations.collection} ${illustrations.dynamicImports.location} ${illustrations.dynamicImports.filterOut.join(" ")}`).join(" && ");
 
 	const postGenerateFolders = illustrationsData.map(illustrations => ({
 		input: illustrations.destinationPath,
@@ -21,7 +21,7 @@ const getScripts = (options) => {
 	}));
 
 	// The script creates the "src/generated/illustrations/**/*.js"
-	const postCreateIllustrationsLoadersScript = postGenerateFolders?.length === 0 ? '' : `node ${LIB}/generate-js-imports/illustrations-root-imports.js '${JSON.stringify(postGenerateFolders)}'`
+	const postCreateIllustrationsLoadersScript = postGenerateFolders?.length === 0 ? '' : `node ${LIB}/generate-js-imports/illustrations.js '${JSON.stringify(postGenerateFolders)}'`
 
 	const tsOption = options.typescript;
 	const tsCommandOld = tsOption ? "tsc" : "";
